+++ CadEditor/settings_flintstones_surprise_at_dinosaur_peak/Settings_Flintstones2-1-rom.cs
@@ -137,11 +137,11 @@ public class Data
     for (int i = 0; i < count; i++)
     {
         var obj = blocksData[i];
-        Globals.romdata[addr + i*4 + 0] = obj.c1;
-        Globals.romdata[addr + i*4 + 2] = obj.c2;
-        Globals.romdata[addr + i*4 + 1] = obj.c3;
-        Globals.romdata[addr + i*4 + 3] = obj.c4;
-        Globals.romdata[addr + count * 4 + i] = obj.typeColor;
+        Globals.romdata[addr + i*4 + 0] = (byte)obj.c1;
+        Globals.romdata[addr + i*4 + 2] = (byte)obj.c2;
+        Globals.romdata[addr + i*4 + 1] = (byte)obj.c3;
+        Globals.romdata[addr + i*4 + 3] = (byte)obj.c4;
+        Globals.romdata[addr + count * 4 + i] = (byte)obj.typeColor;
     }
   }
